requirements.txt
asgiref==3.3.4
Django==3.2
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
django-rest-auth
django-allauth
Pillow
pytest
pytest-django
pytest-cov
mixer
pytest-drf
django-mock-queries
pytest-mock==3.3.1
django-extensions